fixed the dragable tables position on the admin map and server map

+++ web/src/app/pages/admin-map-page/admin-map-page.component.ts
@@ -1,7 +1,8 @@
-import { Component, ViewContainerRef, ViewChild, OnInit, ComponentRef, TemplateRef } from '@angular/core';
+import { Component, ViewContainerRef, ViewChild, OnInit, ComponentRef } from '@angular/core';
 import { InsertionPointDirective } from 'src/app/directives/insertion-point.directive';
 import { DragableIconComponent } from 'src/app/components/dragable-icon/dragable-icon.component';
 import { Table, TableService } from 'src/app/services/table.service';
+import { PlacedDragableIconComponent } from 'src/app/components/placed-dragable-icon/placed-dragable-icon.component';
 
 @Component({
   selector: 'app-admin-map-page',
@@ -11,11 +12,11 @@ import { Table, TableService } from 'src/app/services/table.service';
 export class AdminMapPageComponent implements OnInit {
 
   tableList: ComponentRef<DragableIconComponent>[] = [];
+  placedTableList: ComponentRef<PlacedDragableIconComponent>[] = [];
   tables: Table[];
 
   @ViewChild('tables', { read: ViewContainerRef }) mapContainerRef: ViewContainerRef;
-  @ViewChild('template', { read: TemplateRef }) tableTemplateRef: TemplateRef<ComponentRef<DragableIconComponent>>;
-  @ViewChild(InsertionPointDirective) dragableIcon: InsertionPointDirective;
+  @ViewChild(InsertionPointDirective) insertionPoint: InsertionPointDirective;
 
   constructor(private tableService: TableService) { }
 
@@ -38,11 +39,11 @@ export class AdminMapPageComponent implements OnInit {
   }
 
   createDragableIcon(event: any) {
-    const viewContainerRef = this.dragableIcon.viewContainerRef;
+    const viewContainerRef = this.insertionPoint.viewContainerRef;
 
     const table = {} as Table;
 
-    const componentRef = this.mapContainerRef.createComponent<DragableIconComponent>(DragableIconComponent);
+    const componentRef = viewContainerRef.createComponent<DragableIconComponent>(DragableIconComponent);
     componentRef.instance.table = table;
     componentRef.instance.onDeleteTable.subscribe(tableNum => this.destoyDragableIcon(tableNum));
     this.tableList.push(componentRef);
@@ -50,24 +51,40 @@ export class AdminMapPageComponent implements OnInit {
 
   destoyDragableIcon(deleteNumber: number) {
     const deleteIndex = this.tableList.findIndex(ref => {
-      return ref.instance.table.tableNum === deleteNumber
+      return ref.instance.table.tableNum === deleteNumber;
     });
-    const toDeleteComponent = this.tableList.at(deleteIndex);
-    this.tableList.splice(deleteIndex, 1);
-    toDeleteComponent?.destroy();
 
-    this.tableService.deleteTable(deleteNumber);
+    if (deleteIndex !== -1) {
+      const toDeleteComponent = this.tableList.at(deleteIndex);
+      this.tableList.splice(deleteIndex, 1);
+      toDeleteComponent?.destroy();
+    } else {
+      this.tableService.deleteTable(deleteNumber);
+    }
+
+    const deletePlacedIndex = this.placedTableList.findIndex(ref => {
+      return ref.instance.table.tableNum === deleteNumber;
+    })
+
+    if (deletePlacedIndex !== -1) {
+      const toDeletePlacedComponent = this.placedTableList.at(deletePlacedIndex);
+      this.placedTableList.splice(deletePlacedIndex, 1);
+      toDeletePlacedComponent?.destroy();
+      this.tableService.deleteTable(deleteNumber);
+    }
+
   }
 
   placeTables() {
     this.tables.forEach(table => {
-      let tableRef = this.mapContainerRef.createComponent(DragableIconComponent);
+      let tableRef = this.mapContainerRef.createComponent(PlacedDragableIconComponent);
 
       tableRef.instance.table = table;
       tableRef.instance.dropped = true;
       tableRef.instance.valid = true;
       tableRef.instance.hasValue = true;
-      this.tableList.push(tableRef);
+      tableRef.instance.onDeleteTable.subscribe(tableNum => this.destoyDragableIcon(tableNum));
+      this.placedTableList.push(tableRef);
     })
   }
